backend/requirements.txt
# Backend requirements
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Elasticsearch
elasticsearch==8.11.0

# Data processing
pandas==2.1.4
numpy==1.24.3

# AI dependencies (partagées avec le module AI)
sentence-transformers==2.2.2
torch==2.1.0
transformers==4.35.2

# Utilities
tqdm==4.66.1
requests==2.31.0
